requirements.txt
Flask==3.1.0
pymongo==4.11.3
Flask-Login==0.6.3
APScheduler==3.10.4
requests==2.32.3
cryptography==44.0.0
google-generativeai==0.8.4
openai==1.59.9
anthropic==0.42.0
python-dotenv==1.0.1
